add: gitweb service
Serves all repositories in /data/git on the code. subdomain with gitweb,
and allows cloning them via https:// and git://.

NOTE: For clone over HTTP(S), git update-server-info needs to be run in
the repositories; this is expected to be done via a post-update hook in
the repositories, and not configured here.

Change-Id: I8b5666fd54031447ddfdd098504aa46ad29285b3