add: virtual resource collectors for nginx+pg
This makes it possible to define virtual resources (prefixed by an '@'
character) in other manifests and have them be automatically added to
the correct nodes. For example, a different manifest could define @file
and @front_host resources with tag => [nginx], and the collectors added
here would automatically add them to the front-nginx node.

Change-Id: I8b0fa62ae52632ea228a30bb02fed8bcc2ed8761